# backend-main/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/get-code": {
      "post": {
        "operationId": "AuthController_loginClient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthPhoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully"
          },
          "400": {
            "description": "Invalid phone number"
          }
        },
        "summary": "Request OTP code for login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-code": {
      "post": {
        "operationId": "AuthController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifingCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code verified successfully"
          },
          "400": {
            "description": "Invalid code"
          }
        },
        "summary": "Verify OTP code",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/resend-code": {
      "post": {
        "operationId": "AuthController_resendCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthPhoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code resent successfully"
          },
          "400": {
            "description": "Invalid phone number"
          }
        },
        "summary": "Resend OTP code",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_loginChosenUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginChosenUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login with chosen user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/new-token": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully"
          },
          "400": {
            "description": "Invalid refresh token"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_deconnexion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogOutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "summary": "Logout user",
        "tags": [
          "auth"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Payload invalid"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users found successfully"
          }
        },
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      }
    },
    "/users/phone": {
      "get": {
        "operationId": "UsersController_findByPhone",
        "parameters": [
          {
            "name": "phoneNumber",
            "required": true,
            "in": "query",
            "schema": {
              "example": "0347327950",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users found successfully"
          },
          "400": {
            "description": "Payload invalid"
          }
        },
        "summary": "Find user by phone number",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User found successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Update user by ID",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Delete user by ID",
        "tags": [
          "users"
        ]
      }
    },
    "/register": {
      "post": {
        "operationId": "RegisterController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration created successfully"
          },
          "400": {
            "description": "Invalid data"
          }
        },
        "summary": "Create a new registration",
        "tags": [
          "register"
        ]
      }
    },
    "/formation/{id}": {
      "get": {
        "operationId": "FormationController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Formation found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Formation not found"
          }
        },
        "summary": "Get formation by ID",
        "tags": [
          "Formation"
        ]
      }
    },
    "/sessions": {
      "get": {
        "operationId": "SessionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sessions found successfully"
          }
        },
        "summary": "Get all sessions",
        "tags": [
          "sessions"
        ]
      }
    },
    "/certificats": {
      "get": {
        "operationId": "CertificatsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Certificates retrieved successfully"
          }
        },
        "summary": "Get all certificates",
        "tags": [
          "certificats"
        ]
      }
    },
    "/certificats/{id}": {
      "get": {
        "operationId": "CertificatsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Certificate found"
          },
          "404": {
            "description": "Certificate not found"
          }
        },
        "summary": "Get certificate by ID",
        "tags": [
          "certificats"
        ]
      }
    }
  },
  "info": {
    "title": "Client API",
    "description": "CRUD API for Clients - Backend service for managing client authentication, users, registrations, formations, sessions, and certificates",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "url": "",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "tags": [
    {
      "name": "auth",
      "description": "Authentication endpoints for phone-based login"
    },
    {
      "name": "users",
      "description": "User management endpoints"
    },
    {
      "name": "register",
      "description": "User registration endpoints"
    },
    {
      "name": "Formation",
      "description": "Formation/Training management"
    },
    {
      "name": "Sessions",
      "description": "Session management"
    },
    {
      "name": "Certificats",
      "description": "Certificate management"
    }
  ],
  "servers": [
    {
      "url": "/api",
      "description": "API Base Path"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateAuthPhoneDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "example": "0347327950"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "VerifingCodeDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "example": "0347327950"
          },
          "code": {
            "type": "string",
            "example": "1234"
          }
        },
        "required": [
          "phoneNumber",
          "code"
        ]
      },
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
      "LoginChosenUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "example": "69130451810c61130917a1c6",
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "phone": {
            "type": "string",
            "example": "0347327950"
          },
          "activated": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "phone",
          "activated"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "0bnm-347cvbn-327vbhnj-9bn50"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogOutDto": {
        "type": "object",
        "properties": {
          "id": {
            "example": "0bnm-347cvbn-327vbhnj-9bn50",
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "idUser": {
            "type": "string",
            "description": "Identifiant du customer venant du backend A",
            "example": "42"
          },
          "name": {
            "type": "string",
            "description": "Nom complet de l'utilisateur",
            "example": "Alice Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Numéro de téléphone de l'utilisateur",
            "example": "+261341234567"
          },
          "compteFb": {
            "type": "string",
            "description": "Compte Facebook de l'utilisateur",
            "example": "Miracle Quin"
          }
        },
        "required": [
          "idUser"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "_id": {
            "example": "sdcfvghbnj",
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "idUser": {
            "type": "string",
            "example": "12"
          }
        },
        "required": [
          "_id",
          "idUser"
        ]
      },
      "CreateRegisterDto": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "description": "First name",
            "example": "Alice"
          },
          "lastname": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "sex": {
            "type": "string",
            "description": "Genre l'utilisateur",
            "example": "M"
          },
          "phone": {
            "type": "string",
            "description": "Numéro de téléphone de l'utilisateur",
            "example": "+261341234567"
          },
          "adress": {
            "type": "string",
            "description": "Adresse de l'utilisateur",
            "example": "Fianarantsoa"
          },
          "email": {
            "type": "string",
            "description": "Email de l'utilisateur",
            "example": "jr@rj.j"
          },
          "birthdate": {
            "format": "date-time",
            "type": "string",
            "description": "Date d'anniversaire de l'utilisateur",
            "example": "1997-06-12T21:00:00.000+00:00"
          },
          "birthplace": {
            "type": "string",
            "description": "Lieu de naissance de l'utilisateur",
            "example": "Antananarivo"
          },
          "compteFb": {
            "type": "string",
            "description": "Compte Facebook de l'utilisateur",
            "example": "https://facebook.com/jennie.kim"
          },
          "inactive": {
            "type": "boolean",
            "description": "Compte actif",
            "example": "false"
          }
        },
        "required": [
          "firstname",
          "phone",
          "birthdate",
          "birthplace",
          "inactive"
        ]
      },
      "FormationResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65d0a4bfb663b65ff8d81144"
          },
          "name": {
            "type": "string",
            "example": "Web Fullstack"
          },
          "fullname": {
            "type": "string",
            "example": "Learn modern fullstack development"
          },
          "createdAt": {
            "type": "string",
            "example": "2025-06-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2025-06-01T00:00:00.000Z"
          },
          "inactive": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "_id",
          "name",
          "fullname",
          "createdAt",
          "updatedAt",
          "inactive"
        ]
      }
    }
  }
}